+++ geopdes/src/op_div_v_q.cc
@@ -98,7 +98,8 @@ DEFUN_DLD(op_div_v_q, args, nargout,"OP_DIV_V_Q: assemble the matrix B = [b(i,j)
           } 
         else 
           {
-            warning_with_id ("geopdes:zero_measure_element", "op_div_v_q: element %d has 0 area (or volume)", iel);
+            warning_with_id ("geopdes:zero_measure_element", "op_div_v_q: element %lld has 0 area (or volume)", 
+            static_cast<long long int> (iel));
           }  // end for iel, if area > 0
 
       if (nargout == 1)
