+++ geopdes/src/op_gradu_gradv.cc
@@ -111,7 +111,8 @@ OUTPUT: \n\
                       }  // end for inode, if weight > 0
                 } // end for idof, for jdof
           } else {
-          warning_with_id ("geopdes:zero_measure_element", "op_gradu_gradv: element %d has 0 area (or volume)", iel);
+          warning_with_id ("geopdes:zero_measure_element", "op_gradu_gradv: element %lld has 0 area (or volume)", 
+            static_cast<long long int> (iel));
         } // end for iel, if area > 0
 
       if (nargout == 1)
