+++ geopdes/src/op_su_ev.cc
@@ -152,7 +152,8 @@ DEFUN_DLD(op_su_ev, args, nargout,"\n\
           } 
         else 
           {
-            warning_with_id ("geopdes:zero_measure_element", "op_su_ev: element %d has 0 measure", iel);
+            warning_with_id ("geopdes:zero_measure_element", "op_su_ev: element %lld has 0 measure", 
+            static_cast<long long int> (iel));
           }// end for iel
 
       if (nargout == 1)
